embed lua using provided lua source
Both the source code to lua and a static library are included in this
repo, but neither seems to be specified by the Makefile. Lua files
that define the interpreter are removed because they specify duplicate
definitions of main. The static library is removed because it is no
longer needed.